+++ NEWS.md
@@ -213,6 +213,9 @@ relatively new to CRAN, made these changes sensible at this point.
 
 * `wfm` now uses `mtabulate` and is ~10x faster.
 
+* `plot.polarity` gains arguments for optional error bars using the standard 
+  error of the mean polarity.
+
 <b>CHANGES</b>
 
 * The dictionaries and word lists for <a href="https://github.com/trinker/qdap" target="_blank">qdap</a> have been moved to their own package,